Add build props/targets to WPF, move files to the right location
