+++ samples/WpfBlazorSample/WpfBlazorSample/wwwroot/index.html
@@ -1,0 +1,22 @@
+﻿<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
+    <base href="/" />
+    <link href="css/app.css" rel="stylesheet" />
+    <link href="_content/RazorClassLibrarySample/styles.css" rel="stylesheet" />
+</head>
+
+<body>
+    <app></app>
+
+    <div id="blazor-error-ui">
+        An unhandled error has occurred.
+        <a href="" class="reload">Reload</a>
+        <a class="dismiss">🗙</a>
+    </div>
+</body>
+
+</html>
